rename component. less verbose
